Add input field that edits hero name
Implement two-way binding between the input form element and the
hero.name property using ngModel directive in the component HTML
template.

To do so, import FormsModule in app.module.ts, since NgModel lives
inside FormsModule.

Users are now able to edit the hero name in an <input> textbox.
The textbox displays the hero's name property and update that property
as the user types.